Move user API methods to user_manager
API methods related to the management of users (logging in,
logging out, creating new users, etc.) have been moved to
their own file in the user_manager directory for greater
modularity of components.
